lab1 pass

+++ src/main/java/ru/itmo/tim/controller/WorkerController.java
@@ -72,7 +72,7 @@ public class WorkerController {
         List<String> invalid = new ArrayList<>();
         if (id != null && !id.isEmpty()) {
             try{
-                filters.put("id", Long.parseLong(id));
+                filters.put("id", Integer.parseInt(id));
             }catch(NumberFormatException e){
                 invalid.add(id);
             }
@@ -96,7 +96,6 @@ public class WorkerController {
         if (!invalid.isEmpty()) return Response.status(Response.Status.BAD_REQUEST)
                 .entity(Map.of("error", "Некорректные фильтры", "invalidFields", invalid))
                 .build();
-
         List<WorkerResponseDto> workers = workerService.getAll(page, size, filters, sortColumn, asc);
         Long count = workerService.getCount(filters);
